Reason for Change: fix l1 tests
