+++ test/utests/fakes/FakeABR.cpp
@@ -22,33 +22,29 @@
 long ABRManager::mPersistBandwidth = 0;
 long long ABRManager::mPersistBandwidthUpdatedTime = 0;
 
-ABRManager::ABRManager()
-{
-}
-
 int ABRManager::getProfileCount()
 {
-    return 0;
+	return 0;
 }
 
 int ABRManager::getBestMatchedProfileIndexByBandWidth(int bandwidth)
 {
-    return 0;
+	return 0;
 }
 
 int ABRManager::getMaxBandwidthProfile(const std::string& periodId)
 {
-    return 0;
+	return 0;
 }
 
 long ABRManager::getBandwidthOfProfile(int profileIndex)
 {
-    return 0;
+	return 0;
 }
 
 void ABRManager::clearProfiles()
 {
-    return;
+	return;
 }
 
 void ABRManager::addProfile(ABRManager::ProfileInfo profile)
@@ -57,12 +53,12 @@ void ABRManager::addProfile(ABRManager::ProfileInfo profile)
 
 int ABRManager::getRampedDownProfileIndex(int currentProfileIndex, const std::string& periodId)
 {
-    return 0;
+	return 0;
 }
 
 int ABRManager::getUserDataOfProfile(int currentProfileIndex)
 {
-    return 0;
+	return 0;
 }
 
 void ABRManager::setDefaultInitBitrate(long defaultInitBitrate)
@@ -75,32 +71,32 @@ void ABRManager::updateProfile()
 
 int ABRManager::getDesiredIframeProfile() const
 {
-    return 0;
+	return 0;
 }
 
 int ABRManager::getInitialProfileIndex(bool chooseMediumProfile, const std::string& periodId)
 {
-    return 0;
+	return 0;
 }
 
 int ABRManager::getLowestIframeProfile() const
 {
-    return 0;
+	return 0;
 }
 
 int ABRManager::getProfileIndexByBitrateRampUpOrDown(int currentProfileIndex, long currentBandwidth, long networkBandwidth, int nwConsistencyCnt, const std::string& periodId)
 {
-    return 0;
+	return 0;
 }
 
 int ABRManager::getRampedUpProfileIndex(int currentProfileIndex, const std::string& periodId)
 {
-    return 0;
+	return 0;
 }
 
 bool ABRManager::isProfileIndexBitrateLowest(int currentProfileIndex, const std::string& periodId)
 {
-    return true;
+	return true;
 }
 
 void ABRManager::setDefaultIframeBitrate(long defaultIframeBitrate)
@@ -109,15 +105,97 @@ void ABRManager::setDefaultIframeBitrate(long defaultIframeBitrate)
 
 int ABRManager::removeProfiles(std::vector<BitsPerSecond> profileBPS, int currentProfileIndex, const std::string& periodId)
 {
-    return 0;
+	return 0;
 }
 
 int  ABRManager::getProfileIndexForLowestBandwidth()
 {
-    return 0;
+	return 0;
 }
 
 int ABRManager::getClosestProfileIndexByBandwidth( long inputBandwidth )
 {
-    return 0;
+	return 0;
+}
+
+void ABRManager::ReadPlayerConfig(AampAbrConfig *mAampAbrConfig)
+{
+}
+
+long ABRManager::CheckAbrThresholdSize(int bufferlen, int downloadTimeMs ,long currentProfilebps ,int fragmentDurationMs , CurlAbortReason abortReason)
+{
+	return 0;
+}
+
+void ABRManager::UpdateABRBitrateDataBasedOnCacheLength(std::vector < std::pair<long long,long> > &mAbrBitrateData,long downloadbps,bool LowLatencyMode)
+{
+}
+
+void ABRManager::UpdateABRBitrateDataBasedOnCacheLife(std::vector < std::pair<long long,long> > &mAbrBitrateData , std::vector<BitsPerSecond> &tmpData)
+{
+}
+
+long ABRManager::UpdateABRBitrateDataBasedOnCacheOutlier(std::vector<BitsPerSecond> &tmpData)
+{
+	return 0;
+}
+
+bool ABRManager::CheckProfileChange(double totalFetchedDuration ,int currProfileIndex , long availBW)
+{
+	return false;
+}
+
+bool ABRManager::IsLowestProfile(int currentProfileIndex,bool IstrickplayMode)
+{
+	return false;
+}
+
+void ABRManager::GetDesiredProfileOnBuffer(int currProfileIndex,int &newProfileIndex,double bufferValue,double minBufferNeeded,const std::string& periodId)
+{
+}
+
+
+void ABRManager::CheckRampupFromSteadyState(int currProfileIndex,int &newProfileIndex,long nwBandwidth,double bufferValue,long newBandwidth,BitrateChangeReason &mhBitrateReason,int &mMaxBufferCountCheck,const std::string& periodId)
+{
+}
+
+void ABRManager::CheckRampdownFromSteadyState(int currProfileIndex, int &newProfileIndex,BitrateChangeReason &mBitrateReason,int mABRLowBufferCounter,const std::string& periodId)
+{
+}
+
+long long ABRManager::ABRGetCurrentTimeMS(void)
+{
+	return 0;
+}
+
+bool ABRManager::GetLowLatencyStartABR()
+{
+	return false;
+}
+
+void ABRManager::SetLowLatencyStartABR(bool bStart)
+{
+}
+
+bool ABRManager::GetLowLatencyServiceConfigured()
+{
+	return false;
+}
+
+void ABRManager::SetLowLatencyServiceConfigured(bool bConfig)
+{
+}
+
+bool ABRManager::IsABRDataGoodToEstimate(long time_diff) 
+{
+	return false;
+}
+
+void ABRManager::CheckLLDashABRSpeedStoreSize(struct SpeedCache *speedcache,long &bitsPerSecond,long time_now,long total_dl_diff,long time_diff,long currentTotalDownloaded)
+{
+}
+
+long ABRManager::FragmentfailureRampdown(int buffer,int currentprofileindex)
+{
+	return 0;
 }
